Move decode_response() to response_decoder.py
